Add missing newlines at end of config files
Files without a newline at the end of the file caused issues with the
loading of the next sequential config file if that file's first line was
not a blank line or a comment.  If the first line of the following file
was a keybinding definition for ilia, then the result was that the line
was not parsed properly and ilia could not display the keybinding
information as expected. Ensuring newlines were at the end of all config
files resolved the issue.